// models/DataModals/PatientList.ts
export interface PaitientModel {
  birthday?: number;
  blood?: string;
  email?: string;
  firstName: string;
  gender?: string;
  id: string;
  diseasesCode: string;
  diseasesName: string;
  lastName: string;
  mobilePhone?: string;
  passportNumber?: string;
  status: null;
  ad: string;
  tc?: string;
}